+++ src/frames/context.ts
@@ -15,6 +15,13 @@ export type Context = { [key: string]: Frame };
  */
 export const NilContext: Context = {};
 
+/**
+ * contextEqual compares two contexts for equality.
+ *
+ * @param left
+ * @param right
+ * @returns true if the contexts are equal
+ */
 export function contextEqual(
   left: Context,
   right: Context,
@@ -32,6 +39,12 @@ export function contextEqual(
   return true;
 }
 
+/**
+ * contextString returns a string representation of the context.
+ *
+ * @param context
+ * @returns a string
+ */
 export function contextString(
   context: Context,
 ): string {
@@ -44,7 +57,26 @@ export function contextString(
   }).join(" ");
 }
 
+/**
+ * IKeyValuePair is a tuple of a string and a Frame.
+ */
+/**
+ * Represents a key-value pair where the key is a string and the value is a Frame.
+ * This interface extends ReadonlyArray with the first element being a string (key)
+ * and the second element being a Frame (value).
+ *
+ * @interface IKeyValuePair
+ * @extends {ReadonlyArray<string | Frame>}
+ * @property {string} 0 - The key of the key-value pair.
+ * @property {Frame} 1 - The value of the key-value pair.
+ */
 export interface IKeyValuePair extends ReadonlyArray<string | Frame> {
+  /**
+   * The key of the key-value pair.
+   */
   0: string;
+  /**
+   * The value of the key-value pair.
+   */
   1: Frame;
 }