+++ src/mod.ts
@@ -1,4 +1,14 @@
 export { evaluate } from "./execute/evaluate.ts";
 export { execute } from "./execute/execute.ts";
+export { make_context } from "./execute/hc-eval.ts";
 export { getOptions } from "./cli/hc.ts";
 export { runfile } from "./cli/runfile.ts";
+export { Flatten } from "./cli/flatten.ts";
+export {
+  type Context,
+  type Flags,
+  Frame,
+  type IKeyValuePair,
+  MetaFrame,
+  type StringMap,
+} from "./frames.ts";